create toogler for menu and burger
